requirements.txt
pygame==2.5.0
RPi.GPIO==0.7.1
numpy==1.24.1
pillow==9.4.0
